% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLU.R
\docType{data}
\name{brfss_flu}
\alias{brfss_flu}
\title{BRFSS Place of Influenza Vaccine Data}
\format{
A "long" data frame with XX rows of individual BRFSS respondents in a given state and year and YY column variables:
\describe{
  \item{fluvac_plcat_num}{a numeric variable from the raw BRFSS variable IMFVPLAC}
  \item{fluvac_plcat_fct}{categorical, factor variable of the raw BRFSS variable IMFVPLAC: "Dr. Office","Health Department", "Health Center", "Community Center","Store", "Hospital", "Emergency Room","Workplace", "Other Place", "Other Country","School", "DK/Refused"}
  \item{fluvac_pl5cat_fct}{categorical, factor variable with collapsed categories: Dr. Office, Other Health Setting (Health Dept, Health Center, Hospital, ER), Workplace, Store, or Other (Community Center, Other Place, Other Country, or School). Don't Know or Missing are included in NA.}
  \item{fluvac_allcat_fct}{an extention fluvac_pl5cat_fct that accounts for skip pattern for those without a flu shot: Dr. Office, Other Health Setting, Workplace, Store, Other, Vaccine Place Unkown (got a Flu Shot, but didn't answer IMFVPLAC), No Vaccine (said either No, Don't Know or Refused to answer to Flu Vaccine)}
  \item{x_state}{numeric variable of state}
  \item{state}{state FIPS code, labeled with state alphabetic abbreviation}
  \item{seqno}{identificaiton variable}
  \item{year}{Numeric year: 2016-2017}
  \item{x_psu}{Primary Sampling Unit Variable}
  \item{x_ststr}{Sampling Strata Variable}
  \item{flu_wt_raw}{Original,raw sampling weight: 2016-2017}
  \item{version_flu}{Character, Version of data: Core (X_LLCPWT), V1 (X_LCPWTV1), or V2(X_LCPWTV2)}
}
}
\source{
BRFSS Annual Survey Data \url{https://www.cdc.gov/brfss/annual_data/annual_data.htm}.
}
\usage{
brfss_flu
}
\description{
BRFSS asked the folliwing question from 2016-2017: "At what kind of place did you get your last flu shot or vaccine?" of those who reported getting a flu shot that year.
}
\keyword{datasets}
