% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP.R
\docType{data}
\name{brfss_dep}
\alias{brfss_dep}
\title{BRFSS Depression / Psychological Distress Module Data}
\format{

}
\source{
BRFSS Annual Survey Data \url{https://www.cdc.gov/brfss/annual_data/annual_data.htm}.
}
\usage{
brfss_dep
}
\description{
BRFSS Depression / Psychological Distress Module Questions asked in 2018 are from the PHQ-4 scale
\url{https://www.researchgate.net/publication/40483064_An_Ultra-Brief_Screening_Scale_for_Anxiety_and_Depression_The_PHQ-4}.
}
\examples{
# To Merge depression with core data:
library(tidyverse)
library(brfssR)
data(brfss_dep)
data(brfss_core)
brfss_dep<-full_join(brfss_core,brfss_dep,by="seqno")
}
\keyword{datasets}
